rename contents to sample_tasks
